+++ src/main/java/com/danrus/rpf/compat/rprenames/RpRenamesPlugin.java
@@ -1,0 +1,10 @@
+package com.danrus.rpf.compat.rprenames;
+
+import com.danrus.rpf.compat.RpfCompatPlugin;
+
+public class RpRenamesPlugin extends RpfCompatPlugin {
+    @Override
+    protected String modId() {
+        return "rprenames";
+    }
+}